pyproject.toml
[project]
name = "enwrit"
version = "0.1.0"
description = "Agent instruction management CLI -- compose, port, and score AI agent configs"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
authors = [
    {name = "enwrit", email = "hello@enwrit.com"},
]
keywords = ["ai", "agents", "cli", "instructions", "cursor", "claude", "copilot", "agents-md"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Software Development :: Quality Assurance",
]
dependencies = [
    "typer>=0.15",
    "rich>=13.0",
    "pyyaml>=6.0",
    "pydantic>=2.0",
    "httpx>=0.27",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "ruff>=0.8",
]

[project.urls]
Homepage = "https://enwrit.com"
Repository = "https://github.com/enwrit/writ"
Issues = "https://github.com/enwrit/writ/issues"

[project.scripts]
writ = "writ.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/writ"]

[tool.ruff]
target-version = "py311"
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B"]

[tool.pytest.ini_options]
testpaths = ["tests"]
